Remove LegacyBeatmapConverter.
