+++ osu.Game.Rulesets.Mania/Beatmaps/ManiaBeatmapConverter.cs
@@ -9,6 +9,9 @@ using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Objects.Types;
 using System.Linq;
+using osu.Game.Rulesets.Mania.Beatmaps.Patterns;
+using osu.Game.Rulesets.Mania.MathUtils;
+using osu.Game.Database;
 
 namespace osu.Game.Rulesets.Mania.Beatmaps
 {
@@ -16,19 +19,20 @@ namespace osu.Game.Rulesets.Mania.Beatmaps
     {
         protected override IEnumerable<Type> ValidConversionTypes { get; } = new[] { typeof(IHasXPosition) };
 
+        private Pattern lastPattern = new Pattern();
+        private FastRandom random;
+        private bool isForCurrentRuleset;
+
         protected override Beatmap<ManiaHitObject> ConvertBeatmap(Beatmap original, bool isForCurrentRuleset)
         {
+            this.isForCurrentRuleset = isForCurrentRuleset;
 
-            // Todo: This should be cased when we get better conversion methods
-            var converter = new LegacyBeatmapConverter(original);
+            BeatmapDifficulty difficulty = original.BeatmapInfo.Difficulty;
 
-            return new Beatmap<ManiaHitObject>
-            {
-                BeatmapInfo = original.BeatmapInfo,
-                TimingInfo = original.TimingInfo,
-                // We need to sort here, because the converter generates patterns
-                HitObjects = original.HitObjects.SelectMany(converter.Convert).OrderBy(h => h.StartTime).ToList()
-            };
+            int seed = (int)Math.Round(difficulty.DrainRate + difficulty.CircleSize) * 20 + (int)(difficulty.OverallDifficulty * 41.2) + (int)Math.Round(difficulty.ApproachRate);
+            random = new FastRandom(seed);
+
+            return base.ConvertBeatmap(original, isForCurrentRuleset);
         }
 
         protected override IEnumerable<ManiaHitObject> ConvertHitObject(HitObject original, Beatmap beatmap)
@@ -40,10 +44,102 @@ namespace osu.Game.Rulesets.Mania.Beatmaps
                 yield break;
             }
 
+            IEnumerable<ManiaHitObject> objects;
+            if (isForCurrentRuleset)
+                objects = generateSpecific(original, beatmap);
+            else
+                objects = generateConverted(original, beatmap);
 
+            if (objects == null)
+                yield break;
 
-            // Handled by the LegacyConvereter
-            yield return null;
+            foreach (ManiaHitObject obj in objects)
+                yield return obj;
+        }
+
+        private IEnumerable<ManiaHitObject> generateSpecific(HitObject original, Beatmap beatmap)
+        {
+            var generator = new SpecificBeatmapPatternGenerator(random, original, beatmap, lastPattern);
+
+            Pattern newPattern = generator.Generate();
+            lastPattern = newPattern;
+
+            return newPattern.HitObjects;
+        }
+
+        private IEnumerable<ManiaHitObject> generateConverted(HitObject original, Beatmap beatmap)
+        {
+            var endTimeData = original as IHasEndTime;
+            var distanceData = original as IHasDistance;
+            var positionData = original as IHasPosition;
+
+            // Following lines currently commented out to appease resharper
+
+            //Patterns.PatternGenerator conversion = null;
+
+            if (distanceData != null)
+            {
+                // Slider
+            }
+            else if (endTimeData != null)
+            {
+                // Spinner
+            }
+            else if (positionData != null)
+            {
+                // Circle
+            }
+
+            //if (conversion == null)
+            return null;
+
+            //Pattern newPattern = conversion.Generate();
+            //lastPattern = newPattern;
+
+            //return newPattern.HitObjects;
+        }
+
+        /// <summary>
+        /// A pattern generator for mania-specific beatmaps.
+        /// </summary>
+        private class SpecificBeatmapPatternGenerator : Patterns.Legacy.PatternGenerator
+        {
+            public SpecificBeatmapPatternGenerator(FastRandom random, HitObject hitObject, Beatmap beatmap, Pattern previousPattern)
+                : base(random, hitObject, beatmap, previousPattern)
+            {
+            }
+
+            public override Pattern Generate()
+            {
+                var endTimeData = HitObject as IHasEndTime;
+                var positionData = HitObject as IHasXPosition;
+
+                int column = GetColumn(positionData?.X ?? 0);
+
+                var pattern = new Pattern();
+
+                if (endTimeData != null)
+                {
+                    pattern.Add(new HoldNote
+                    {
+                        StartTime = HitObject.StartTime,
+                        Samples = HitObject.Samples,
+                        Duration = endTimeData.Duration,
+                        Column = column,
+                    });
+                }
+                else if (positionData != null)
+                {
+                    pattern.Add(new Note
+                    {
+                        StartTime = HitObject.StartTime,
+                        Samples = HitObject.Samples,
+                        Column = column
+                    });
+                }
+
+                return pattern;
+            }
         }
     }
 }
