+++ osu.Game.Rulesets.Mania/osu.Game.Rulesets.Mania.csproj
@@ -49,7 +49,6 @@
   <ItemGroup>
     <Compile Include="Beatmaps\Patterns\Legacy\PatternGenerator.cs" />
     <Compile Include="Beatmaps\Patterns\PatternGenerator.cs" />
-    <Compile Include="Beatmaps\LegacyBeatmapConverter.cs" />
     <Compile Include="Beatmaps\Patterns\Legacy\PatternType.cs" />
     <Compile Include="Beatmaps\ManiaBeatmapConverter.cs" />
     <Compile Include="Beatmaps\Patterns\Pattern.cs" />
